Refactor save payment method js logic
